Streamline CI workflow
This package has no complicated code or dependencies hence is does
not makes sense to explicitly test against Windows and Mac
environments. This is simply saves time, resources and energy :)

Add support for Go 19.x and keep supporting 16.x for some more
time as it was a major milestone too.